Removed google protobuf
Added game state player iterator, as well datum iterator.
Maybe fixed a crash with Himanshu's tag loader.
Disabled UPNP for dedicated servers.
Removed some duplicate code.
Ported Hitfix code to use tag interface code, same for FOV slider and crosshair position.
H2X firerates enabled by default in Campaign.
Added some player data structures.
Removed unneeded memory allocations for object placement data new.
Added configurable LIVE netcode values (TODO: reverse what those values do)
Hooked the real map load function instead of a random function, which also gets called not only on map load.
Cleaned header includes.
And some other misc changes I forgot to mention.

TODOS:
Fix AdvLobbySettings, as well map checksum sync to use the new custom packets.

+++ xlive/H2MOD/Modules/Tweaks/Tweaks.h
@@ -17,8 +17,8 @@ namespace H2Tweaks {
 	void applyShaderTweaks();
 	void enable60FPSCutscenes();
 	void disable60FPSCutscenes();
-	void setFOV(double field_of_view_degrees);
-	void setVehicleFOV(double field_of_view_degrees);
+	void setFOV(int field_of_view_degrees);
+	void setVehicleFOV(int field_of_view_degrees);
 	void setHz();
 	void setCrosshairPos(float crosshair_offset);
 	void setCrosshairSize(int size, bool preset);
